requirements.txt
amqp==2.1.1
audioread==2.1.5
billiard==3.5.0.3
celery==4.0.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==2.0
django-celery-results==1.0.1
django-classy-tags==0.8.0
django-cors-headers==2.1.0
django-sekizai==0.10.0
djangorestframework==3.7.1
drf-nested-routers==0.90.0
idna==2.6
influxdb==4.1.1
joblib==0.11
kombu==4.0.0
librosa==0.5.1
llvmlite==0.20.0
neo4j-driver==1.5.0
numba==0.35.0
numpy==1.13.3
polyglotdb==0.1.13
py==1.4.34
pyraat==0.3.1
pytest==3.2.3
pytest-django==3.1.2
pytest-pythonpath==0.7.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
resampy==0.2.0
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
TextGrid==1.4
tqdm==4.19.4
urllib3==1.22
vine==1.1.4
